fixing refs, and clearfilter, and incorporated columnFormatter function
