+++ modules/help_editor/jsx/help_editor.js
@@ -1,5 +1,4 @@
 import FilterForm from 'FilterForm';
-import formatColumn from './columnFormatter';
 
 /**
  * Help Editor Archive Page.
@@ -62,7 +61,7 @@ class HelpEditor extends React.Component {
   }
 
   resetFilters() {
-    this.refs.helpFilter.clearFilter();
+    this.filter.clearFilter();
   }
 
   render() {
@@ -81,7 +80,6 @@ class HelpEditor extends React.Component {
     const filterRef = function(f) {
       this.filter = f;
     }.bind(this);
-
     return (
       <div>
         <FilterForm
@@ -116,8 +114,51 @@ HelpEditor.propTypes = {
   DataURL: React.PropTypes.string.isRequired
 };
 
+/* exported formatColumn */
+
 /**
- * Render dicom_page on page load
+ * Modify behaviour of specified column cells in the Data Table component
+ * @param {string} column - column name
+ * @param {string} cell - cell content
+ * @param {arrray} rowData - array of cell contents for a specific row
+ * @param {arrray} rowHeaders - array of table headers (column names)
+ * @return {*} a formated table cell for a given column
+ */
+function formatColumn(column, cell, rowData, rowHeaders) {
+  if (loris.hiddenHeaders.indexOf(column) > -1) {
+    return null;
+  }
+  // Create the mapping between rowHeaders and rowData in a row object.
+  var row = {};
+  var url;
+  rowHeaders.forEach(function(header, index) {
+    row[header] = rowData[index];
+  }, this);
+
+  if (column === 'Topic') {
+    url = loris.BaseURL + "/help_editor/edit_help_content/?helpID=" +
+           row.HelpID + "&parentID=" + row.ParentID;
+    return <td>
+                <a href ={url}>{cell}</a>
+             </td>;
+  }
+  if (column === 'Parent Topic') {
+    url = loris.BaseURL + "/help_editor/edit_help_content/?helpID=" +
+           row.ParentID + "&parentID=" + row.ParentTopicID;
+    return <td>
+                <a href ={url}>{cell}</a>
+             </td>;
+  }
+
+  return <td>{cell}</td>;
+}
+
+window.formatColumn = formatColumn;
+
+export default formatColumn;
+
+/**
+ * Render help_editor on page load
  */
 window.onload = function() {
   var dataURL = loris.BaseURL + "/help_editor/?format=json";
